% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellspecR-package.R
\docType{package}
\name{cellspecR-package}
\alias{cellspecR}
\alias{cellspecR-package}
\title{cellspecR: Read, Validate and Store Multiplexed Imaging Cell Tables}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

cellspecR defines \code{cellspec}, one table format for segmented cells from
multiplexed tissue images, reads the exports of common image-analysis tools
into it, checks the result, and writes it to disk with a checksum manifest
so that later analysis steps can prove what they read.

The format is described in \code{vignette("specification", package = "cellspecR")}.
}
\section{Main functions}{

\itemize{
\item Specification: \code{\link[=cs_spec_version]{cs_spec_version()}}, \code{\link[=cs_vocabulary]{cs_vocabulary()}}.
\item Object: \code{\link[=cs_new]{cs_new()}}, \code{\link[=cs_cells]{cs_cells()}}, \code{\link[=cs_measurements]{cs_measurements()}}, \code{\link[=cs_dictionary]{cs_dictionary()}},
\code{\link[=cs_images]{cs_images()}}, \code{\link[=cs_channels]{cs_channels()}}, \code{\link[=cs_provenance]{cs_provenance()}}, \code{\link[=cs_adjacency]{cs_adjacency()}},
\code{\link[=cs_features]{cs_features()}}, \code{\link[=cs_markers]{cs_markers()}}, \code{\link[=is_cellspec]{is_cellspec()}}.
\item Methods: \code{\link[=print.cellspec]{print.cellspec()}}, \code{\link[=summary.cellspec]{summary.cellspec()}}, \code{\link[=dim.cellspec]{dim.cellspec()}},
\code{[.cellspec}, \code{\link[=as.data.frame.cellspec]{as.data.frame.cellspec()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/cellspecR}
  \item \url{https://cttir.github.io/cellspecR/}
  \item Report bugs at \url{https://github.com/CTTIR/cellspecR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]
}

}
\keyword{internal}
