% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cs_bind}
\alias{cs_bind}
\title{Combine cellspec objects from different images}
\usage{
cs_bind(...)
}
\arguments{
\item{...}{\code{cellspec} objects to combine.}
}
\value{
A combined \code{cellspec} object.
}
\description{
\code{cs_bind()} combines cell tables, measurements and metadata from one or
more \code{cellspec} objects. Objects must describe different images and must
use the same measurement dictionary.
}
\examples{
cs_bind(cs_example())
}
\seealso{
Other object:
\code{\link{cs_accessors}},
\code{\link[=cs_example]{cs_example()}},
\code{\link[=cs_features]{cs_features()}},
\code{\link[=cs_merge_tiles]{cs_merge_tiles()}},
\code{\link[=cs_new]{cs_new()}},
\code{\link[=is_cellspec]{is_cellspec()}}
}
\concept{object}
