% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cs_merge_tiles}
\alias{cs_merge_tiles}
\title{Merge tiled cellspec objects using centroid ownership}
\usage{
cs_merge_tiles(tiles, tile_bounds, ownership = "centroid", tolerance = 1e-09)
}
\arguments{
\item{tiles}{A named or positional list of \code{cellspec} tile objects. Names,
when supplied, must equal \code{tile_bounds$tile_id}.}

\item{tile_bounds}{A data frame with \code{tile_id}, \code{xmin}, \code{xmax}, \code{ymin} and
\code{ymax} in micrometres.}

\item{ownership}{Ownership method. The only supported value is
\code{"centroid"}.}

\item{tolerance}{Non-negative tolerance used to snap coordinates close to a
box boundary before applying its half open comparisons.}
}
\value{
A merged \code{cellspec} object.
}
\description{
Every centroid is assigned to the half open outer box containing it. A
centroid found in more than one box is an error. Cells in a tile whose
centroid is owned by another tile, or by no tile, are treated as halo cells
and dropped.
}
\examples{
tile_bounds <- data.frame(
  tile_id = "tile-1", xmin = 0, xmax = 10000, ymin = 0, ymax = 10000
)
cs_merge_tiles(list(cs_example()), tile_bounds)
}
\seealso{
Other object:
\code{\link{cs_accessors}},
\code{\link[=cs_bind]{cs_bind()}},
\code{\link[=cs_example]{cs_example()}},
\code{\link[=cs_features]{cs_features()}},
\code{\link[=cs_new]{cs_new()}},
\code{\link[=is_cellspec]{is_cellspec()}}
}
\concept{object}
