% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is_cellspec}
\alias{is_cellspec}
\title{Test for a cellspec object}
\usage{
is_cellspec(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"cellspec"}, otherwise \code{FALSE}. The
object is not validated; use \code{\link[=cs_validate]{cs_validate()}} for that.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
is_cellspec(cs_example())
is_cellspec(data.frame())
}
\seealso{
Other object:
\code{\link{cs_accessors}},
\code{\link[=cs_bind]{cs_bind()}},
\code{\link[=cs_example]{cs_example()}},
\code{\link[=cs_features]{cs_features()}},
\code{\link[=cs_merge_tiles]{cs_merge_tiles()}},
\code{\link[=cs_new]{cs_new()}}
}
\concept{object}
